LayoutTests/canvas/philip/tests/2d.imageData.put.unchanged-expected.txt

< [index] >
2d.imageData.put.unchanged
putImageData(getImageData(...), ...) has no effect
References: 2d.imageData.unchanged
Actual output:
Passed
